% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_fulltext_to_xml.R
\name{pdf_fulltext_to_xml}
\alias{pdf_fulltext_to_xml}
\title{Process PDFs with API}
\usage{
pdf_fulltext_to_xml(
  input,
  output = getwd(),
  api_url,
  tei_coordinates = c("figure", "biblStruct"),
  force = FALSE
)
}
\arguments{
\item{input}{The path to a PDF file or a folder containing PDF files to be processed.}

\item{output}{The folder where the processed TEI XML files will be saved. Default is the current working directory.}

\item{api_url}{The URL of the API to process the PDF files.}

\item{tei_coordinates}{A character vector specifying the TEI coordinates to be included in the API request. Default is c("figure", "biblStruct").}

\item{force}{Logical, if TRUE, existing output files will be overwritten. Default is FALSE.}
}
\value{
invisible(NULL)
}
\description{
This function processes PDF files using an API and saves the results as TEI XML files in the specified output folder.
}
\examples{
\dontrun{
process_with_api("path/to/input/folder", "path/to/output/folder", "http://example.com/api")
}
}
