% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_tibble.R
\name{xml_to_tibble}
\alias{xml_to_tibble}
\title{Convert XML files to a tibble}
\usage{
xml_to_tibble(input_dir)
}
\arguments{
\item{input_dir}{A character string specifying the directory containing the XML files.}
}
\value{
A tibble with columns `title`, `doi`, and `body`, where each row corresponds
to an XML file.
}
\description{
This function reads all XML files with a `.tei.xml` extension from a specified
directory, extracts the title, DOI, and body text from each file, and combines
the results into a tibble.
}
